Derive Eq on compatible types.

+++ tests/schemas/multi_valued_union_records.rs
@@ -1,10 +1,10 @@
 
 #[derive(Debug, PartialEq, Clone, serde::Deserialize, serde::Serialize)]
 pub struct Baz {
-    pub d: i32,
+    pub d: f32,
 }
 
-#[derive(Debug, PartialEq, Clone, serde::Deserialize, serde::Serialize)]
+#[derive(Debug, PartialEq, Eq, Clone, serde::Deserialize, serde::Serialize)]
 pub struct Bar {
     pub c: i64,
 }
